// evaluador-front/src/app/core/models/ingreso.model.ts
import { Avaluo } from './avaluo.model';
import { Inspeccion } from './inspeccion.model';

export interface Ingresos {
  id: number;
  tiposervicio: string;
  solicitante: string;
  documento_solicitante: string;
  direccion_solicitante: string;
  telefono_solicitante: string;
  placa: string;
  ubicacion_activo: string;
  fecha_solicitud: string;
  fecha_inspeccion: string;
  fecha_informe: string;
  objeto_avaluo: string;
  codigo_interno_movil: string;
  tipo_propiedad: string;
  fecha_matricula: string;
  movil: string;
  marca: string;
  linea: string;
  clase: string;
  tipo_carroceria: string;
  cantidad_ejes:number;
  categoria: string;
  color: string;
  cilindraje: number;
  modelo: number;
  kilometraje: number;
  caja_cambios: string;
  tipo_traccion: string;
  numero_pasajeros: number;
  capacidad_carga: number;
  llanta_delantera_izquierda: string | null;
  llanta_delantera_derecha: string | null;
  llanta_trasera_izquierda: string | null;
  llanta_trasera_derecha: string | null;
  llanta_repuesto: string | null;
  numero_chasis: string;
  numero_serie: string;
  numero_motor: string;
  nacionalidad: string;
  propietario: string;
  documento_propietario:string;
  empresa_afiliacion: string;
  soat: string;
  fecha_expedicion_soat: string;
  fecha_inicio_vigencia_soat: string;
  fecha_vencimiento_soat: string;
  entidad_expide_soat: string;
  estado_soat: string;
  rtm: string;
  fecha_vencimiento_rtm: string;
  centro_revision_rtm: string;
  estado_rtm: string;
  ciudad_registro: string;
  no_licencia: string;
  fecha_expedicion_licencia: string;
  organismo_transito: string;
  estado: string;
  created_at: string;
  updated_at: string;
  avaluo: Avaluo;
  inspeccion:Inspeccion;
  peso_mermado:string;
}